Refactor: Move types for query-builder to src/types
