+++ impl/c-qube/src/types/QueryBuilderSchema.ts
